--- entry/card.js
@@ -1,39 +1,0 @@
-export default [
-  [
-    {},
-    {
-      label: '订单编号',
-      value: '20180912123045',
-    },
-    {
-      label: '时间',
-      value: '2018-09-12 12:30:45',
-    },
-  ],
-  [
-    {
-      label: '金额',
-      value: 'V-T To justify printed text means to adjust the spaces between the words so that each line of type is exactly the same length. 使 (文本) 对齐',
-    },
-    {},
-    {
-      value: '一个商品',
-    },
-  ],
-  [
-    {
-      label: '金额',
-      align: 100,
-      value: '2018-09-12 12:30:45',
-    },
-    {
-      label: '商品',
-      value: '一个商品',
-    },
-  ],
-  [
-    {
-      value: 'aksdj4',
-    },
-  ],
-]
