+++ entry/index.html
@@ -2,7 +2,7 @@
 <html>
 <head>
 <title> owl </title>
-<link rel="stylesheet" href="//unpkg.alipay.com/antd/dist/antd.min.css" />
+<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/antd/3.11.2/antd.min.css" />
 </head>
 <body>
   <div id="root"></div>
