+++ entry/index.js
@@ -1,13 +1,9 @@
 import React from 'react'
 import { render } from 'react-dom'
 import Tag from 'antd/lib/tag'
-import Entry from '../src'
-import Card from '../src/card'
-import { data, formater } from './data'
-import cardData from './card'
-
+import { Cell, Card } from '../src'
+import { cellData, cardData, formater } from './data'
 import '../src/index.less'
-import '../src/card.less'
 
 const customs = {
   tag({ text, props }) {
@@ -17,10 +13,9 @@ const customs = {
   },
 }
 
-// render(<Entry
-//   customs={customs}
-//   data={data}
-//   formater={formater}
-// />, document.getElementById('root'))
-
-render(<Card align="justify" data={cardData} />, document.getElementById('root'))
+render((
+  <div>
+    <Cell customs={customs} data={cellData} formater={formater} />
+    <Card align="justify" data={cardData} />
+  </div>
+), document.getElementById('root'))
