+++ entry/umd.js
@@ -1,4 +1,4 @@
-import Owl from '../src'
+import * as Owl from '../src'
 import '../src/index.less'
 
 export default Owl
