+++ src/index.js
@@ -1,12 +1,2 @@
-import React from 'react'
-import parser from './parser'
-
-export default function ({
-  formater = {},
-  data = {},
-  customs = {},
-  className = 'react-owl',
-}) {
-  const P = parser(formater, customs, className)
-  return (<P {...data} />)
-}
+export { default as Cell } from './cell'
+export { default as Card } from './card'
